ARCEMEAnalysis

Documentation for ARCEMEAnalysis.

ARCEMEAnalysis.Event

ARCEMEAnalysis._arceme_cubenames

ARCEMEAnalysis.arceme_all_fingerprints

ARCEMEAnalysis.arceme_bias_corrected_fp

ARCEMEAnalysis.arceme_coordinates

ARCEMEAnalysis.arceme_create_indexcubes

ARCEMEAnalysis.arceme_endtime

ARCEMEAnalysis.arceme_eventdate

ARCEMEAnalysis.arceme_eventlist

ARCEMEAnalysis.arceme_eventpairs

ARCEMEAnalysis.arceme_kndvi

ARCEMEAnalysis.arceme_landcover

ARCEMEAnalysis.arceme_merge_fingerprints

ARCEMEAnalysis.arceme_ndvi

ARCEMEAnalysis.arceme_open

ARCEMEAnalysis.arceme_optical_band_fingerprints

ARCEMEAnalysis.arceme_radar_fingerprints

ARCEMEAnalysis.arceme_rgb

ARCEMEAnalysis.arceme_s1_position

ARCEMEAnalysis.arceme_set_httpstore

ARCEMEAnalysis.arceme_set_localpath

ARCEMEAnalysis.arceme_spectral

ARCEMEAnalysis.arceme_starttime

ARCEMEAnalysis.arceme_uncorrected_fp

ARCEMEAnalysis.arceme_validpairs

ARCEMEAnalysis.boa

ARCEMEAnalysis.Event — Type

struct Event

Define a structure to hold information about an ARCEME event.

ARCEMEAnalysis._arceme_cubenames — Method

_arceme_cubenames(;batch="6")

List all available data cube names in the specified batch stored in the ARCEME S3 bucket.

ARCEMEAnalysis.arceme_all_fingerprints — Method

arceme_all_fingerprints(ds_d, ds_dhp)

Computes both optical and radar fingerprints for the ARCEME data cube datasets ds_d and ds_dhp. Also aggregates the fingerprints over time in 2-monthly windows for the whol 2 years around the event date.

ARCEMEAnalysis.arceme_bias_corrected_fp — Function

arceme_bias_corrected_fp(inputcube,cloudcube,lccube)

Computes a cloud-biased corrected footprint aggregated by land cover class for the provided inputcube, cloud mask and land cover cube.

ARCEMEAnalysis.arceme_coordinates — Method

arceme_coordinates(ds)

Get the longitude and latitude coordinates of the center of the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_create_indexcubes — Method

For every valid event pair creates a and stores data cubes of a list of precomputed vegetation indices. For kNDVI a shared sigma parameter per land cover class is computed.

ARCEMEAnalysis.arceme_endtime — Method

arceme_endtime(ds)

Get the end time of the ARCEME data cube dataset ds

ARCEMEAnalysis.arceme_eventdate — Method

arceme_eventdate(ds)

Get the time of when the event in an ARCEME data cube dataset ds happened.

ARCEMEAnalysis.arceme_eventlist — Method

arceme_eventlist()

Get a list of ARCEME events from the local CSV file.

ARCEMEAnalysis.arceme_eventpairs — Method

arceme_eventpairs()

Get pairs of ARCEME events from the event list.

ARCEMEAnalysis.arceme_kndvi — Function

ARCEMEAnalysis.arceme_landcover — Method

 arceme_landcover(ds)

Get the land cover statistics for the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_merge_fingerprints — Method

arceme_merge_fingerprints(optical_fp, radar_fp)

Merges the optical and radar fingerprints datasets into a single dataset.

ARCEMEAnalysis.arceme_ndvi — Method

arceme_ndvi(ds)

Compute the NDVI (Normalized Difference Vegetation Index) for the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_open — Method

arceme_open(cubename; batch="ARCEME-DC-6")

Open the specified ARCEME data cube from the local path (if with arcemesetlocalpath) or the httpstore (default "https://s3.waw3-2.cloudferro.com/swift/v1", reset with arcemesethttpstore).

ARCEMEAnalysis.arceme_optical_band_fingerprints — Method

arceme_optical_band_fingerprints(ds_d, ds_dhp)

Computes fingerprints for all optical bands in the ARCEME data cube datasets ds_d and ds_dhp. Also aggregates the fingerprints over time in 2-monthly windows for the whol 2 years around the event date.

ARCEMEAnalysis.arceme_radar_fingerprints — Method

arceme_radar_fingerprints(ds_d, ds_dhp)

Computes fingerprints for all radar bands in the ARCEME data cube datasets ds_d and ds_dhp. Also aggregates the fingerprints over time in 2-monthly windows for the whol 2 years around the event date.

ARCEMEAnalysis.arceme_rgb — Method

arceme_rgb(ds)

Compute the RGB composite for the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_s1_position — Method

arceme_s1_position(ds)

Analyses the sentinel 1 time stamps of the datasets and groups them into series of time stamps whose difference is always a multiple of 6 days, so we can assume that all images with the same group tag are retrieved from the same position. Store the groups in the dataset with the name s1_postion.

ARCEMEAnalysis.arceme_set_httpstore — Method

arcemesethttpstore(httpstore)

Sets the https store from which to retrieve the cubes. If no local preference is set, the default location is used: https://s3.waw3-2.cloudferro.com/swift/v1

Setting this preference needs recompilation.

ARCEMEAnalysis.arceme_set_localpath — Method

arceme_set_localpath(path)

Sets the path to a local directory where downloaded copies of the arceme cubes are stored.

ARCEMEAnalysis.arceme_spectral — Method

arceme_spectral(ds, indices::Vector{String})

Compute the listed indices using SpectralIndices.jl. Not working.

ARCEMEAnalysis.arceme_starttime — Method

arceme_starttime(ds)

Get the start time of the ARCEME data cube dataset ds

ARCEMEAnalysis.arceme_uncorrected_fp — Method

arceme_uncorrected_fp(inputcube,dataset; timeaxis=:time_sentinel_1_rtc)

Computes a cloud-biased corrected footprint aggregated by land cover class for the provided inputcube, cloud mask and land cover cube.

ARCEMEAnalysis.arceme_validpairs — Method

arceme_validpairs(;batch="ARCEME-DC-6")

Get valid pairs of ARCEME events from the local path (if set with arcemesetlocalpath) or the http data store. Default httpstore is "https://s3.waw3-2.cloudferro.com/swift/v1". It can be reset with arcemesethttpstore.

ARCEMEAnalysis.boa — Method

boa(band; BOA_ADD_OFFSET = -1000, QUANTIFICATION_VALUE = 10000)

Compute the radiometric offsets for Sentinel 2 bands to get values at bottom of atmosphere. The default values are valid for data processed from baseline 04.00 (January 2022) onwards. The entire Sentinel-2 archive in CDSE has been reprocessed and is now available in baseline 05.xx, with consistent offset.