ARCEMEAnalysis

Documentation for ARCEMEAnalysis.

ARCEMEAnalysis.Event

ARCEMEAnalysis._arceme_cubenames

ARCEMEAnalysis.arceme_bias_corrected_fp

ARCEMEAnalysis.arceme_coordinates

ARCEMEAnalysis.arceme_endtime

ARCEMEAnalysis.arceme_eventdate

ARCEMEAnalysis.arceme_eventlist

ARCEMEAnalysis.arceme_eventpairs

ARCEMEAnalysis.arceme_landcover

ARCEMEAnalysis.arceme_ndvi

ARCEMEAnalysis.arceme_open

ARCEMEAnalysis.arceme_rgb

ARCEMEAnalysis.arceme_starttime

ARCEMEAnalysis.Event — Type

struct Event

Define a structure to hold information about an ARCEME event.

ARCEMEAnalysis._arceme_cubenames — Method

_arceme_cubenames(;batch="SECONDBATCH")

List all available data cube names in the specified batch stored in the ARCEME S3 bucket.

ARCEMEAnalysis.arceme_bias_corrected_fp — Method

arceme_bias_corrected_fp(inputcube,cloudcube,lccube)

Computes a cloud-biased corrected footprint aggregated by land cover class for the provided inputcube, cloud mask and land cover cube.

ARCEMEAnalysis.arceme_coordinates — Method

arceme_coordinates(ds)

Get the longitude and latitude coordinates of the center of the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_endtime — Method

arceme_endtime(ds)

Get the end time of the ARCEME data cube dataset ds

ARCEMEAnalysis.arceme_eventdate — Method

arceme_eventdate(ds)

Get the time of when the event in an ARCEME data cube dataset ds happened.

ARCEMEAnalysis.arceme_eventlist — Method

arceme_eventlist()

Get a list of ARCEME events from the local CSV file.

ARCEMEAnalysis.arceme_eventpairs — Method

arceme_eventpairs()

Get pairs of ARCEME events from the event list.

ARCEMEAnalysis.arceme_landcover — Method

 arceme_landcover(ds)

Get the land cover statistics for the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_ndvi — Method

arceme_ndvi(ds)

Compute the NDVI (Normalized Difference Vegetation Index) for the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_open — Method

arceme_open(cubename; batch="SECONDBATCH")

Open the specified ARCEME data cube from the S3 bucket.

ARCEMEAnalysis.arceme_rgb — Method

arceme_rgb(ds)

Compute the RGB composite for the ARCEME data cube dataset ds.

ARCEMEAnalysis.arceme_starttime — Method

arceme_starttime(ds)

Get the start time of the ARCEME data cube dataset ds